requirements.txt
Django>=3.2.5,<3.3
psycopg2>=2.8.6,<2.9
Pillow>=9.2.0,<9.3
import-export==0.2.67.dev6
django-import-export #==2.7.1
graphene>=2.1.9,<3.0
djangorestframework>=3.13.1,<3.14
graphene-django>=2.15.0,<2.16
graphene-file-upload>=1.3.0,<1.4